# swin_transformer_v2.py error  RuntimeError fixed